fixes for; 1. results and summary of results 2. solo stackers from connected account do btc transfer
