+++ src/lib/components/all-voters/VoteResultsRow.svelte
@@ -5,7 +5,6 @@ import cross from '$lib/assets/cross.png'
 	import NakamotoResultsBackground from '../shared/NakamotoResultsBackground.svelte';
 	import { fmtMicroToStx, fmtNumber } from '$lib/utils';
 
-export let stxPower: number;
 export let accountsFor = 0;
 export let accountsAgainst = 0;
 export let stxFor = 0;
@@ -16,7 +15,7 @@ export let stxAgainst = 0;
 <div class="mb-8 lg:grid lg:gap-8 lg:grid-cols-3 space-y-4 lg:space-y-0">
 	<div class="p-8 bg-[#121314] rounded-2xl items-stretch">
 		<h3 class="mb-8 text-3xl text-white">Voting power</h3>
-		<h3 class="text-1xl text-white">{fmtNumber(Math.floor(Number(fmtMicroToStx(stxPower))))} stacks voted</h3>
+		<h3 class="text-1xl text-white">{fmtNumber(Math.floor(Number(fmtMicroToStx(stxFor + stxAgainst))))} stacks voted</h3>
 	</div>
 	<div class="p-8 bg-primary-01 rounded-2xl relative">
 		<div class="flex justify-between mb-5">
